Improved AcquisitionWorker error handling and timing
Initial testing shows this solves the position boxes not updating issues.
Added exception handling to AcquisitionWorker's run loop to log errors and emit a signal with None data on failure to keep the loop going. Changed data acquisition interval to 200 ms and added a 2-second timeout to the wait condition.